+++ rails/app/views/home/top.html.slim
@@ -1,4 +1,4 @@
-.ui.inverted.vertical.masthead.center.aligned.segment#top-image
+.ui.inverted.vertical.masthead.center.aligned.segment#top-image data-controller='anime'
   .ui.text.container
     h1.ui.header.top-title
       |Coffee-oma
